+++ packages/react/src/components/Grid/Grid.types.ts
@@ -1,0 +1,4 @@
+import { ComponentProps } from '../../types/ComponentProps'
+import { Grid } from './Grid'
+
+export type GridProps = ComponentProps<typeof Grid>