+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @luizlahr-ui/tokens
 
+## 1.1.1
+
+### Patch Changes
+
+- - Add FormControl component
+  - Add isInvalid props to TextInput
+
 ## 1.1.0
 
 ### Minor Changes